Add meeting minutes for the TSC meeting on 2025-01-17
The purpose of this meeting was to discuss the proposal to move the ORT
project from the Linux Foundation (LF) to the Eclipse Foundation (EF),
see [1].

The Markdown was created by exporting the Google Drive document
created by Google Gemini without any modifications. In order to maintain
the exact document created by Gemini, any corrections should be done as
separate commits on top.

[1]: https://github.com/oss-review-toolkit/ort-governance/issues/35

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>